+++ src/sampling/triangulate.py
@@ -2,7 +2,7 @@ import numpy as np
 import cv2
 from typing import List, Tuple
 
-from util.geometrytypes import PointList, TriangleList
+from util.geometry_types import PointList, TriangleList
 from util.settings import Settings
 
 class Triangulate:
